gorsh. add rds and shit connection string implementation. separated infra by responsibilities
